Fix: #40. Use key file path instead of key file names
  - In the AWSAdapter, currently the config file takes in the key file name
    and the adapter code assumes the key file is the current directory and
assumes it is a .pem file(attaches a .pem extension to the end of the key file
name). This should be changed to have the entire key file path in the config
file. Then the adapter can be agnostic about the type and location of the key
file. Fix this.